+++ src/Sharers/GooglePlusSharer.php
@@ -17,6 +17,8 @@ use MichaelHall\LinkSharer\Sharers\Base\AbstractSharer;
 /**
  * Google Plus sharer.
  *
+ * @deprecated
+ *
  * @since 1.0.0
  */
 class GooglePlusSharer extends AbstractSharer